# swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string",
						"nullable": true
					},
					"data": {
						"additionalProperties": true,
						"type": "object"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Client": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"age": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"zip_code": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"vehicules": {
						"items": {
							"$ref": "#/components/schemas/Vehicule"
						},
						"type": "array",
						"nullable": true
					},
					"compagnies": {
						"items": {
							"$ref": "#/components/schemas/Company"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"firstName",
					"lastName",
					"phone",
					"address",
					"city",
					"zip_code",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Vehicule": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"marque": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"pictures": {
						"type": "string",
						"nullable": true
					},
					"createdDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"client": {
						"$ref": "#/components/schemas/Client"
					}
				},
				"required": [
					"marque",
					"model",
					"color",
					"client"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Company": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"zip_code": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array",
						"nullable": true
					},
					"clients": {
						"items": {
							"$ref": "#/components/schemas/Client"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"name",
					"phone",
					"address",
					"city",
					"zip_code",
					"country"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"birthdate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"zip_code": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"picture": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"company": {
						"$ref": "#/components/schemas/Company",
						"nullable": true
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"phone",
					"address",
					"city",
					"zip_code",
					"country",
					"role",
					"password",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Settings": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"notification": {
						"type": "boolean"
					},
					"theme": {
						"type": "boolean"
					},
					"sound": {
						"type": "boolean"
					},
					"createdDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"notification",
					"theme",
					"sound",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "authorization"
			}
		}
	},
	"info": {
		"title": "backendicm",
		"version": "0.0.0",
		"license": {
			"name": "MIT"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/clients/get-all": {
			"get": {
				"operationId": "GetClients",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Client"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/v1/clients/get/{id}": {
			"get": {
				"operationId": "GetClient",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/clients/create": {
			"post": {
				"operationId": "CreateClient",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Client"
							}
						}
					}
				}
			}
		},
		"/v1/clients/update/{id}": {
			"put": {
				"operationId": "UpdateClient",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Client"
							}
						}
					}
				}
			}
		},
		"/v1/clients/delete/{id}": {
			"delete": {
				"operationId": "DeleteClient",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/compagnies/get/{id}": {
			"get": {
				"operationId": "GetCompany",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/compagnies/create": {
			"post": {
				"operationId": "CreateCompany",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Company"
							}
						}
					}
				}
			}
		},
		"/v1/compagnies/update/{id}": {
			"put": {
				"operationId": "UpdateCompany",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Company"
							}
						}
					}
				}
			}
		},
		"/v1/compagnies/delete/{id}": {
			"delete": {
				"operationId": "DeleteCompany",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Company"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/settings/get/{id}": {
			"get": {
				"operationId": "GetSettings",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/settings/update/{id}": {
			"put": {
				"operationId": "UpdateSettings",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Settings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Settings"
							}
						}
					}
				}
			}
		},
...
		"/v1/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/v1/auth/signin": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:3000/"
		}
	]
}